// frontend/src/common/enums/app/app-route.enum.ts
enum AppRoute {
  ROOT = '/',
  SIGN_UP_AS = '/sign-up-as',
  SIGN_UP_PATIENT = '/sign-up-patient',
  SIGN_UP_PRACTITIONER = '/sign-up-practitioner',
  SIGN_IN = '/sign-in',
  ACTIVATION_LINK_SENT = '/activation-link-sent',
  CREATE_NEW_PASSWORD = '/create-new-password',
  FORGOT_PASSWORD = '/forgot-password',
  CHECK_EMAIL = '/check-email',
  PATIENT = '/patient',
  PATIENT_QUESTIONARY_INFO = '/patient/questionary/info',
  PATIENT_QUESTIONARY_VERIFICATION = '/patient/questionary/verification',
  PATIENT_QUESTIONARY_BMI = '/patient/questionary/bmi',
  PATIENT_QUESTIONARY_MEDICAL_INFO = '/patient/questionary/medical-info',
  PATIENT_QUESTIONARY_CONTACT_PERSON_INFO = '/patient/questionary/contact-person-info',
  PRACTITIONER_QUESTIONNAIRE_INFO = '/practitioner/questionnaire/info',
  PRACTITIONER_QUESTIONNAIRE_VERIFICATION = '/practitioner/questionnaire/verification',
  PRACTITIONER_QUESTIONNAIRE_EDUCATION = '/practitioner/questionnaire/education',
  PRACTITIONER_QUESTIONNAIRE_WORK_EXPERIENCE = '/practitioner/questionnaire/work-experience',
  PRACTITIONER_PORTAL = '/portal/practitioner',
  PRACTITIONER_PORTAL_DECLARATIONS = '/portal/practitioner/declaration-requests',
  PRACTITIONER_PORTAL_PROFILE = '/portal/practitioner/profile',
  PRACTITIONER_PORTAL_PROFILE_INFO = '/portal/practitioner/profile/personal-information',
  PRACTITIONER_PORTAL_PROFILE_PASSWORD = '/portal/practitioner/profile/change-password',
  PRACTITIONER_PORTAL_PROFILE_EDUCATION = '/portal/practitioner/profile/education',
  PRACTITIONER_PORTAL_PROFILE_EXPERIENCE = '/portal/practitioner/profile/work-experience',
  PRACTITIONER_PORTAL_PROFILE_DOCUMENTS = '/portal/practitioner/profile/documents',
  PRACTITIONER_PORTAL_PROFILE_NOTIFICATIONS = '/portal/practitioner/profile/notifications',
  PRACTITIONER_PORTAL_PROFILE_DELETE = '/portal/practitioner/profile/delete-profile',
  PRACTITIONER_PORTAL_PATIENTS = '/portal/practitioner/my-patients',
  PRACTITIONER_PORTAL_PATIENT_DETAILS = '/portal/practitioner/patient/details',
  PRACTITIONER_PORTAL_VISITS = '/portal/practitioner/patient-visits',
  PRACTITIONER_PORTAL_SCHEDULE = '/portal/practitioner/schedule',
  PRACTITIONER_PORTAL_CHAT = '/portal/practitioner/chat',
  PRACTITIONER_QUESTIONARY_INFO = '/practitioner/questionary/info',
  PRACTITIONER_QUESTIONARY_VERIFICATION = '/practitioner/questionary/verification',
  PRACTITIONER_QUESTIONARY_PROFESSIONAL_INFO = '/practitioner/questionary/professional-info',
  PATIENT_PORTAL = '/portal/patient',
  PATIENT_PORTAL_PRACTITIONERS = '/portal/patient/my-practitioners',
  PATIENT_PORTAL_PRACTITIONERS_PROFILE = '/portal/patient/my-practitioners/profile',
  PATIENT_PORTAL_MY_PRACTITIONER = '/portal/patient/my-practitioner',
  PATIENT_PORTAL_SCHEDULE = '/portal/patient/my-practitioner/schedule',
  PATIENT_PORTAL_SCHEDULE_APPOINTMENT = '/portal/patient/my-practitioner/make-appointment',
  PATIENT_PORTAL_HEALTH_HISTORY = '/portal/patient/health-history',
  PATIENT_PORTAL_INSURANCE = '/portal/patient/insurance',
  PATIENT_PORTAL_RELATED_PEOPLE = '/portal/patient/related-people',
  PATIENT_PORTAL_ANALYTICS = '/portal/patient/analytics',
  PATIENT_PORTAL_PROFILE = '/portal/patient/profile',
  PATIENT_PORTAL_PROFILE_INFO = '/portal/patient/profile/personal-information',
  PATIENT_PORTAL_PROFILE_PASSWORD = '/portal/patient/profile/change-password',
  PATIENT_PORTAL_PROFILE_BMI = '/portal/patient/profile/bmi',
  PATIENT_PORTAL_PROFILE_MEDICAL_INFO = '/portal/patient/profile/medical-info',
  PATIENT_PORTAL_PROFILE_CONTACT_PEOPLE = '/portal/patient/profile/contact-people',
  PATIENT_PORTAL_PROFILE_DOCUMENTS = '/portal/patient/profile/documents',
  PATIENT_PORTAL_PROFILE_NOTIFICATIONS = '/portal/patient/profile/notifications',
  PATIENT_PORTAL_PROFILE_DELETE = '/portal/patient/profile/delete-profile',
  PATIENT_PORTAL_VACCINES = '/portal/patient/vaccines',
  PATIENT_PORTAL_CHAT = '/portal/patient/chat',
}

export { AppRoute };